Move PackageName stuff out of package.dart.
These classes are largely independent of the Package class.
