+++ lib/src/solver/unselected_package_queue.dart
@@ -8,7 +8,7 @@ import 'dart:collection';
 import 'package:stack_trace/stack_trace.dart';
 
 import '../log.dart' as log;
-import '../package.dart';
+import '../package_name.dart';
 import 'backtracking_solver.dart';
 
 /// A priority queue of package references.
