+++ lib/src/source/hosted.dart
@@ -16,6 +16,7 @@ import '../http.dart';
 import '../io.dart';
 import '../log.dart' as log;
 import '../package.dart';
+import '../package_name.dart';
 import '../pubspec.dart';
 import '../source.dart';
 import '../system_cache.dart';
